fix(types): mypy --strict 全量清零（30 项，含 0.16.3 起存量 29 项）
CI quality job 从 0.16.3 起连红（先 ruff 后 mypy 从未跑达）。清单：
queue_protocol 15（6 缺参数注解/detail/tags 两次 .get narrow 失效/
_member_meta cache 可空/display 重名/死代码 by_ref 行）；server 7
（fake-mcp 双路径 CallToolResult=None 赋值/别名重导入/4 处 wrapper
no-any-return 用 cast——注解不可动，MCP output schema 从注解派生）；
scan_pipeline 5（outcome dict 显式 Any/内联 set 参数化）；其余 3
（cosine float 收窄/truncation 分支 early_result 改名/hit_spans 注解）。
mypy+ruff+1411 tests+slow 11 全绿。

+++ memory_arbiter/difference_classifier.py
@@ -122,7 +122,7 @@ def name_cosine(a: str, b: str) -> float:
     set_a, set_b = _bigrams(a), _bigrams(b)
     if not set_a or not set_b:
         return 0.0
-    return len(set_a & set_b) / ((len(set_a) * len(set_b)) ** 0.5)
+    return float(len(set_a & set_b)) / float((len(set_a) * len(set_b)) ** 0.5)
 
 
 def _tokens(text: str) -> set[str]:
